requirements.txt
Django
Pillow
PyJWT
asgiref
cachetools
certifi
cffi
charset-normalizer
cryptography
defusedxml
django-allauth
django-recaptcha
google-api-core
google-api-python-client
google-auth
google-auth-httplib2
googleapis-common-protos
httplib2
idna
oauthlib
pip
protobuf
psycopg2
pyasn1
pyasn1-modules
pycparser
pyparsing
python3-openid
requests
requests-oauthlib
rsa
setuptools
six
sqlparse
uritemplate
urllib3
wheel
flake8